Isolating job calculation logic for analysis
